use pretrained multi lang embeddings

+++ vico/scripts/download.py
@@ -17,18 +17,25 @@ def format_file_name(url):
 
 def download_all(label_path, output_folder):
     label_frame = pandas.read_csv(label_path)
-    for url in label_frame.product_page_url:
-        download(url, output_folder)
+    paths = []
+    for url in label_frame.url:
+        path = download(url, output_folder)
+        paths.append(path)
+    label_frame['path'] = paths
+    label_frame.to_csv(label_path, index=False)
 
 
 def download(url, output_folder):
     file_name = format_file_name(url)
     file_path = os.path.join(output_folder, file_name)
+    if os.path.isfile(file_path):
+        return file_path
     response = requests.get(url)
     if response.status_code != 200:
-        return
+        return None
     with open(file_path, 'w') as f:
         f.write(response.text)
+    return file_path
 
 
 def parse_args():
